+++ Examples/checkDataForSubject.py
@@ -18,8 +18,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 
 
-This script checks if all videos are uploaded for a subject across several 
-sessions.
+This script checks if all videos are uploaded and not error for a subject 
+across however many sessions are associated with them.
 
 
 You will need login credentials from app.opencap.ai to run this script. You 
@@ -29,16 +29,11 @@ prior to running this script. Otherwise, you will need to to login every time
 you run the script.
 """
 
-import requests
 import sys
 import os
 sys.path.append(os.path.abspath('./..'))
 
-from utilsAuth import getToken
-from utilsAPI import getAPIURL
-
-API_URL = getAPIURL()
-API_TOKEN = getToken()
+import utils
 
 # %% Editable variables
 
@@ -51,69 +46,15 @@ trialNames = [
               ['stairs_up','stairs_down']
             ]
 
-# %% functions
 
-def getSubjectNumber(subjectName):
-    subjects = requests.get(API_URL + "subjects/",
-                           headers = {"Authorization": "Token {}".format(API_TOKEN)}).json()
-    sNum = [s['id'] for s in subjects if s['name'] == subjectName]
-    if len(sNum)>1:
-        print(len(sNum) + ' subjects with the name ' + subjectName + '.')   
-    elif len(sNum) == 0:
-        print('no subject found with this name.')
-        
-    return sNum[0]
-
-def getUserSessions():
-    sessionJson = requests.get(API_URL + "sessions/valid/",
-                           headers = {"Authorization": "Token {}".format(API_TOKEN)}).json()
-    return sessionJson
-
-def getParticipantSessions(subjectName):
-    sessions = getUserSessions()
-    subNum = getSubjectNumber(subjectName)
-    sessions2 = [s for s in sessions if (s['subject'] == subNum)]
-    
-    return sessions2
-
-def getTrialNames(session):
-    trialNames = [t['name'] for t in session['trials']]
-    return trialNames
-
-def findSessionWithTrials(subjectTrialNames,trialNames):
-    hasTrials = []
-    for trials in trialNames:
-        hasTrials.append(None)
-        for i,sTrials in enumerate(subjectTrialNames):
-            if all(elem in sTrials for elem in trials):
-                hasTrials[-1] = i
-                break
-            
-    return hasTrials
-
-def get_entry_with_largest_number(lst):
-    max_entry = None
-    max_number = float('-inf')
-
-    for entry in lst:
-        # Extract the number from the string
-        try:
-            number = int(entry.split('_')[-1])
-            if number > max_number:
-                max_number = number
-                max_entry = entry
-        except ValueError:
-            continue
-
-    return max_entry
 # %% main
 
-subjectSessions = getParticipantSessions(subjectName)
+subjectSessions = utils.getParticipantSessions(subjectName)
 if len(subjectSessions) == 0:
     raise Exception('no sessions for this subject')
     
-subjectTrialNames = [getTrialNames(s) for s in subjectSessions]
-sessionMapping = findSessionWithTrials(subjectTrialNames,trialNames)
+subjectTrialNames = [utils.getTrialNames(s) for s in subjectSessions]
+sessionMapping = utils.findSessionWithTrials(subjectTrialNames,trialNames)
 
 for i,sMap in enumerate(sessionMapping):
     if sMap is None:
@@ -126,7 +67,7 @@ for i,sMap in enumerate(sessionMapping):
     for t in trialNames[i]:
         trials = [st for st in subjectTrialNamesThisSession if t in st]
         if len(trials) > 1:
-            goodTrialNames.append(get_entry_with_largest_number(trials))
+            goodTrialNames.append(utils.get_entry_with_largest_number(trials))
         else:
             goodTrialNames.append(trials[0])
